requirements.txt
fastapi>=0.68.0
uvicorn[standard]
streamlit
python-dotenv
twilio
